Version 1.5.2 - Tested on Rebot.
The big problem is that any use of the telemetry.AddData function from the Robot class crashes the running program. I needed to gut the ReportStatus method and return to using minimal telemetry call from with the Op mode.

After re-re-adjusting motor directions and the Drive|Strafe|Turn "matrix", The robot drive as expected.

Shoot motors are spinning as expected, but accordong to their lables the motors are mounted on the wrong sides. If the hardare were correted, then the shooter would be functioning as expected.